JJB 2.0.X does not allow scenario to be passed here
With this change releng/jjb/* now passes on:
Jenkins Job Builder version: 2.0.1 and 1.6.2

Error:
TypeError: write() argument must be str, not list
Explanation:
in jjb 1.6.2 it would allow the list in and then the job would convert
the list into a literal string of "['var1', 'var2']"

- 'compass-{scenario}-{pod}-daily-{stream}'
Can have scenario passed, {scenario} is part of the job name
so it knows which scenario to add.

- 'compass-deploy-{pod}-daily-{stream}'
cannot have scenario passed, there is not one job generated per scenario
so its tryting to add "['senario1', 'scenario2', 'etc']" into the {scenario}
value, which is not allowed.

This worked on jjb 1.6.2 becasue we overwrite the scenario
information later:
- trigger-builds:
    - project: 'compass-deploy-{pod}-daily-{stream}'
      predefined-parameters: |
        DEPLOY_SCENARIO={scenario}

jjb/armband/armband-ci-jobs.yml is very simillar to this job
and passes a default, which works, so that is what I have done here:
   - string:
       name: DEPLOY_SCENARIO
       default: 'os-odl-nofeature-ha'

addendum:
removed scenario: '{scenario}' from
'compass-{scenario}-{pod}-daily-{stream}'
even though it is valid, Trevor rightly pointed out that there is
nothing to consume this value.

Change-Id: I134b2bd5888dcdfa5b768a7db2bdc98ad32ec28e
Signed-off-by: agardner <agardner@linuxfoundation.org>
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>