JJB 2.0.X does not allow scenario to be passed here
With this change releng/jjb/* now passes on:
Jenkins Job Builder version: 2.0.1

Error:
TypeError: write() argument must be str, not list
Explanation:
in jjb 1.6.2 it would allow the list in and then the job would convert
the list into a literal string of "['var1', 'var2']"

I think we have this information in the var
DEPLOY_SCENARIO and hopefully dont need it in parameters:

eg:
can't use this:
 scenario: '{scenario}'

can use this:
- string:
    name: DEPLOY_SCENARIO
    default: '{scenario}'

jjb/armband/armband-ci-jobs.yml is very simillar to this job
and passes

   - string:
       name: DEPLOY_SCENARIO
       default: 'os-odl-nofeature-ha'

as a paramerter.
perhaps there is a reason for this, and that it will help.

Change-Id: I134b2bd5888dcdfa5b768a7db2bdc98ad32ec28e
Signed-off-by: agardner <agardner@linuxfoundation.org>