openci: Add OPNFV scenario composition and test jobs for the demo
The flow is
- OPNFV compose job gets triggered by the ArtifactPublishedEvents published by
ODL and ONAP autorelease jobs, composing the corresponding scenarios. At the
end, CompositionDefinedEvent is published for the composed scenario.
- OPNFV test job gets triggered by the CompositionDefinedEvents published by
OPNFV compose jobs, deploying and testing the composed scenario. At the end,
ConfidenceLevelModifiedEvent is published for the tested scenario which can
be consumed by ODL and/or ONAP Jenkinses, doing further promotion.

Change-Id: Ice1736d65f3f2b26e849302fec2bb0c579da4a1c
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>