Update Jenkins Job Builder and Global-JJB
Updates Jenkins Job Builder to 2.0.7 and Global-JJB 0.19.2.

These need to be done in unison due to the update to Postbuildscript:
Postbuildscript 2.X isn't supported until Jenkins Job Builder 2.0.3, and
Global-JJB didn't support the new Postbuildscript formatting until
0.18.0.

This brings in changes which fix issues seen in Apex and XCI when using
Mulitjob phases along with adding support for new plugins such as
jms-messaging.

The Gerrit change for the Global-JJB submodule commit can be found here:
https://gerrit.linuxfoundation.org/infra/#/c/10536/

Change-Id: If3e48a07ccee14c60f552a7ed26481a7640185ba
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>