openci: Add ODL autorelease and promotion jobs for the demo
The jjb will be proposed to ODL Releng once it is verified on OPNFV
Jenkins.

The flow is
- ODL autorelease job publishes an ArtifactPublishedEvent to make the
availability of a new version that is tested known to the world
- ODL promote job gets triggered by a ConfidenceLevelModifiedEvent
published by OPNFV CI to promote autorelease artifact further.

The prototype doing realistic stuff will be built on top of these jobs
and
- parse the triggering event to find the location of autorelease
artifact
- download the artifact for composing ODL scenario
- run the deployment and testing of ODL scenario

Please note that this change just adds the basic structure for the jobs
to share the idea. Further changes will add triggering, message
costruction, and publishing.

Change-Id: I51d7e0cfcb5f2552fcc9ac287ed12ed70f582895
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>