openci: Add ONAP autorelease and promotion jobs for the demo
The jjb will be proposed to ONAP Releng once it is verified on OPNFV
Jenkins.

The flow is
- ONAP autorelease job publishes an ArtifactPublishedEvent to make the
availability of a new version that is tested known to the world
- ONAP promote job gets triggered by a ConfidenceLevelModifiedEvent
published by OPNFV CI to promote autorelease artifact further.

The prototype doing realistic stuff will be built on top of these jobs
and
- parse the triggering event to find the location of autorelease
artifact
- download the artifact for composing ONAP scenario
- run the deployment and testing of ONAP scenario

Please note that this change just adds the basic structure for the jobs
to share the idea. Further changes will add triggering, message
costruction, and publishing.

Change-Id: Ib8939ab44d664ff71f472db66a0e2171ed7b56c5
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>