Update Global-JJB to master
This is a temporary update until a new release of Global-JJB comes out.

Updating Global-JJB to a commit on master past 0.19.2 which allows
jjb jobs to run on the same builder without tripping over each other.

The related Gerrit change for the current tip of master can be seen
here: https://gerrit.linuxfoundation.org/infra/#/c/10717/

Change-Id: I4205fc74d8af805ac29ac93d4c5b033c145a00b9
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>