Update patch set 2
Patch Set 2:

Hi,
To get IPMI creds from encrypted PDF, there are at least 2 ways I can think about:
- decrypt the PDF using the key on the jumpserver (we already do this as part of the FUel install using Pharos's utils/generate_config.py);
- read the pillar data from Salt master using salt calls;
The advantage of the first method is that it is not installer specific, but requires integrating with the Pharos generate_config.py.
The second approach is simpler (just query the Salt master with something like `salt mas\* pillar.get maas.machines` - writing this from memory, the actual parameter might be slightly different), but it is specific to Fuel.

Patch-set: 2