A fix to the error of using x86 shared object file
Currently, our Yardstick job on arm64(aarch64) platform fails
due to a error usage of the x86 shared object file libzmq.so.3
which is located in:
/opt/nsb_bin/trex/scripts/external_libs/pyzmq-14.5.0/python2
/ucs4/64bit/zmq/libzmq.so.3
,I replaced with that of arm64 platform and it can work now.

This fix would check and replace the wrong shared object file
with the correct one if any. So it would be still useful for
jobs when using some certain release of Yardstick.

I have raised a Jira issue in OPNFV about the problem:
 https://jira.opnfv.org/browse/YARDSTICK-1191

Change-Id: I50e80bb89e51039ebce66d412bdc6634cf316773
Signed-off-by: trevortao <trevor.tao@arm.com>