Projects did not fill out their releases/fraser/
This is the patchset of projects that did not fill out their
releases fraser file.

I have done it for them.
some of the projects had no changes on the fraser branch
in this case I put the tag in the same place as the 6.0.0 tag

some project I am familliar with and I move the tag to the
head of fraser for them,

Ideally team leads will do this work in the future
We will update the tagging instructions accordingly.

Change-Id: I1b2c7a8b6e63bc857192757e25f2d103c2629319
Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>