Update patch set 1
Patch Set 1:

> Sorry for the flip flop, but given: https://gerrit.opnfv.org/gerrit/#/c/57725/
 > looks about ready to merge, is there a reason this would take
 > precedence?
 > 
 > Glad you found a temporary fix, but if the other change is merged,
 > this would have to be reverted anyways...

Hi Trevor:-)
Thank you very much for your review and comments.
The current fix is compatible with the change of

 > Sorry for the flip flop, but given: https://gerrit.opnfv.org/gerrit/#/c/57725/
 > looks about ready to merge, is there a reason this would take
 > precedence?
 > 
 > Glad you found a temporary fix, but if the other change is merged,
 > this would have to be reverted anyways...

Hi Trevor:-)
Thank you very much for the review and comments.
This fix is compatible with the change of
https://gerrit.opnfv.org/gerrit/#/c/57725/ as it would check whether the old style of code distribution(v2.28) of TRex is used, if yes, it would replace it with correct shared object file.
And it would fix the error of using x86 shared object file for the release before the change of '57725' is merged, such as Fraser release.
So could you please have a new look at it?

Patch-set: 1