Add separate fraser jobs to Barometer container builds
In order to support additonal features that have been added to
collectd since the 5.8 release, there needs to be a separate
stable and master container so we can support the new features
as well as the stable release.

This change will allow us to increment the version of collectd
as we add functest support, config guides and documentation,
while keeping the stable version container intact and available
for users to download.

JIRA: BAROMETER-102

Change-Id: I06ca84581a2e0f1a994cedf53aebc8f5cc3e7535
Signed-off-by: Emma Foley <emma.l.foley@intel.com>