Update patch set 1
Patch Set 1: Code-Review-1

So as the code is right now, our CI wouldn't execute the scenario for k8s unless you started it with a gate job.  In that gate job, we do not use apex-virtual, we call deploy and test explicitly:
https://github.com/opnfv/releng/blob/master/jjb/apex/apex-verify-jobs.yaml#L255

So you would need to disable it there.

Patch-set: 1
Label: Code-Review=-1