xci: Disable Centos jobs until Intel lab is back
Due to Intel lab maintenance, the number of slaves will go down so we need
to have fewer jobs running. Centos is known to be problematic upstream and
it is better to disable them until the lab is back online to ensure the jobs
for other distros are impacted less.

Change-Id: Ifd8624e45282d68bd7acdf283ef7a499e81990e9
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>