Update patch set 1
Patch Set 1:

Here is the link form Sanbox that verifies that the Job runs.

https://sandbox.opnfv.org/view/All/job/sfc-verify-master/9/console

The reason that is failing is not because the job is misconfigured but because we have a bug in SFC code with some broken URLs. As you can see from the above link the rest tox test are running without a problem.

Patch-set: 1