Apex: Fix OS credentials fetch
Now that keystone V3 Identity is only one supported, OOO 'overcloudrc'
should be used in Queens and later as that contains V3 creds.
'overcloudrc.v3' will eventually be removed in OOO.

Note, this was causing tempest_smoke_serial to fail in Apex with Swift
test cases. This is due to the fact that in Apex we were only setting
the OS_REGION_NAME in overcloudrc, and functest was using
overcloudrc.v3. When Tempest does not have the region set it will simply
look for the first endpoint it can find for that service type, and was
not using the right keystone endpoint for Apex.

Change-Id: I154c9e762dfc5ba035f6e108606ecedcfb0d60d2
Signed-off-by: Tim Rozet <trozet@redhat.com>