Update patch set 1
Patch Set 1:

I tested Thahn's patch [0] which is fixing the problem but to consume it we need to wiat first for the guys to merge the patch and then do a BUmp SHA to the OPNFV releng repo so we can use the new patch of the global-jjb whcih fixes the problem.

We can wait until then and do the appropriate changes to releng and SFC code or we merge this patch as it is and we create new patches to cinsume the tox-verify-job when evreything is in place. 

What do you think guys ?

Is up to you.

[0] https://gerrit.linuxfoundation.org/infra/11513

Patch-set: 1
Reviewer: Gerrit User 6745 <6745@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=0