Functest: Restores ability to run with pre-defined creds
This patch is needed for the case where there are provided OS
credentials to functest, and there is no need to find the installer VM
and copy the credentials to the host. In the Apex snapshot case, there
is no installer VM on the slave, so we must provide the credentials
file.

Also includes a fix to use tempest_smoke_serial for the snapshot netvirt
verification until CSIT suite is working correctly.

Change-Id: Ic84955f527e3f84bb0d8579a377b4a1883fe3eb2
Signed-off-by: Tim Rozet <trozet@redhat.com>