Disable SFC verify job in fraser branch
We would like to disable the SFC verify job for fraser branch
because jenkins give us -1 on new patches on stable fraser branch
because of the lack of tox on that branch.

We are not going to submit any new patches on the SFC fraser branch from
now on except from one [0]. So we can disable the fraser job as an easy
fix to the -1 jenkins-ci problem in gerrit.

[0]: https://gerrit.opnfv.org/gerrit/#/c/59175/

Change-Id: Iab2cb1eea6ac14f262ca25bc3cc61e15556f21e9
Signed-off-by: Dimitrios Markou <mardim@intracom-telecom.com>