[dovetail-webportal] Add docker build jobs for dovetail-webportal
Add 2 docker build jobs for the new repo dovetail-webportal.
One is dovetail-webportal-api, the other is dovetail-webportal-web.
Once there is a new patch merged, it will auto trigger the jobs to build
and upload the new docker images with tag 'latest'.
According to the Dovetail release plan and develpoment of the web, can tag
this repo and release with certain tag (e.g. ovp.2.0.0)

JIRA: DOVETAIL-671

Change-Id: I1becc084617bb52f8c8e23275202bd7be4a7b2b3
Signed-off-by: xudan <xudan16@huawei.com>