Ensure docker prunning can be done
Sometimes the docker build fails and the docker process hangs on a container,
withouth being able to stop it. This causes spurious failures on other builds
that shouldn't fail. The prune job should cleanup the hanging containers and
images, but it fails when the docker process is hanged.

This change adds a docker restart to the prune job, in order to assure that the
hanged container is shut down and can be removed.

JIRA: ARMBAND-358

Change-Id: I9d04fa1099a6fc53bf773a2eba33be89db7d0aa1
Signed-off-by: Cristina Pauna <cristina.pauna@enea.com>