Apex: Allow for  multi-version snapshots
This patch adds ability to manage multiple snapshot versions. We want to
be able to have snapshots for noha and ha for multiple OS releases. This
adds that capability and updates the deploy script to be able to deploy
them. Also a user can specify which scenario to trigger with his ODL
netvirt commit by doing 'opnfv-check <os version>-<ha|noha>

Change-Id: I3884b41c86b55c187becfb83ef62f3431724d09e
Signed-off-by: Tim Rozet <trozet@redhat.com>