Update patch set 2
Patch Set 2:

> (1 comment)

Yes, sure. My only concern was the following: upon each run, Dovetail cleans the results directory, but the result package is stored outside of the results directory and is not deleted. So, after running Dovetail multiple times, there are multiple result file packages in $DOVETAIL_HOME. Now I was wondering if those file also accumulate on the CI machine so that for each CI run, we also upload the old result files or if the CI machinery cleans the entire workspace after each run.

Patch-set: 2