xci: Reconfigure xci-daily-virtual jobs
The execute shell parts need to be switched to existing
scripts but since the scripts were mainly written for
verify jobs, the adaptations they require has potential
to cause issues so it needs to be done in a separate
change.

Apart from that, the verify and daily jobs can't be run
on same slave at the moment due to a bug in opnfv vm
creation playbook. If clean vm exists on the node, the
playbook fails. Build blocker is enable for verify and
daily jobs to ensure the jobs don't end up on same slave
for the timebeing.

Change-Id: Iefaf5a5c2d4a79bd48a5f9da28d68748bc67f511
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>