Update patch set 2
Patch Set 2:

releng-jjb-verify and releng-jjb-merge jobs started failing for some issue with lftools so waiting for Trevor/Aric to look at that since I am not sure what would go wrong with the job updates if we submit this.

See an example failure on below build. It could be harmless but I don't know.
https://build.opnfv.org/ci/job/releng-jjb-merge/278/console

Patch-set: 2