jjb: xci: Do not wipe clean VM on failures
If there is a CI failure, then we need the VM to be present so we can
ssh to the slave and collect all the logs we can in order to debug
the problem. However, in order to not leave stale VMs around, we should
remove the VM if a job is scheduled again on the same node. As such, if
we need to debug a build on the node, we need to disconnect the node
from Jenkins so the VM is not wiped by a future job.

Change-Id: I93e7d7402fe03c392b63b87e7d438191f6e7e9cc
Signed-off-by: Markos Chandras <mchandras@suse.de>