Rework creating tags & branches from release files
Old method could not properly handle both branching and
tagging on changes to the release file.

Proposed method:
jjb calls:
jjb/releng/branch-or-tag.sh
for both verify and merge jobs.

branch-or-tag determins nature of the change to
"releases/branch/project.yaml"
and calls
 releng-release-tagging.sh
or
 releng-release-create-branch.sh

The scripts above handle both verify and merge operations

based on
  if [[ $TAG_EXISTS = false && "$JOB_NAME" =~ "merge" ]];
and
  if [[ $REF_EXISTS = true && "$JOB_NAME" =~ "merge" ]]; then

both scripts check that the ref exits in verfiy and merge stages.

if releng-release-create-branch.sh creates a branch it then also
calls releases/scripts/create_jobs.py

Change-Id: Ieb99de5e4df100af59ecc818f52b0831383b2b62
Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>