Update patch set 2
Patch Set 2: Code-Review-1

(3 comments)

I'm not sure breaking the code up into functions that get ran serially makes 
much sense (one calling the next).

Overall here's what the release automation should be doing.

For any changes to the release files we want to verify:
 - The schema is correct (-1/+1)

Verify the following for branching: 
 - The ref is on master and the branch does not exist (-1/+1)
 - Skip if branch exists (+1)

Verify the following for tagging:
 - The ref for the tag exists and is on the stable branch (-1/+1)
 - The ref for the stable branch is the current stream
   [ex: stable/fraser for fraser jobs]
 - Skip if tag exists (+1)

Merging for branches:
 - Create the branch (skip if exists)
 - Create patches for Jenkins jobs (script will skip of jobs exist)

Merging for tags:
 - Create the tag (skip if exists)

I'm not sure if we need to do all the verification again for merges, but it 
would be simple to have a JOB_NAME =~ merge check for the branches and tags
scripts.

I think we can have the tagging in one file and branching logic in another along with combining the verification and merge jobs into those file, so both jobs would have the same builder:
 - create-venv.sh
 - tagging.sh
 - branching.sh

Patch-set: 2
Label: Code-Review=-1