Update patch set 1
Patch Set 1: Code-Review+2

> > The run of this job in the sandbox failed, but I've yet to verify
 > > if the job is also failing (and in the same way) in production.
 > 
 > do you have some logs or a job link? I wonder what went wrong

Ah, the job can't actually be ran/tested unless it's against a patchset from a project, and there is no parameter to specify GERRIT_PROJECT to do that.

The doc job succeeded but the publishing at the end failed, which is what caused the failure.

Patch-set: 1
Label: Code-Review=+2