Update patch set 3
Patch Set 3:

> This needs to be rebased on top of: 63047

OK, I must admit I wrote that https://gerrit.opnfv.org/gerrit/63047 is a prerequisite for this change.
I meant 63047 is needed to get the full effect of this change on L18-L22 of yardstick-daily-jobs.yaml, and is required for the PODs to alias the &gambia anchor, which has not been done at this stage.
If possible, I would prefer that this change is not further delayed, while waiting for 63047.
L27 of yardstick-daily-jobs.yaml can be fixed with 'ovp-2.0.0' docker tag.
Most important is yardstick-project-jobs.yaml, which will enable the yardstick-verify-gambia and yardstick-merge-gambia jobs for Jenkins.

Patch-set: 3