CPERF: Do not remove docker images in clean
Removing the cperf docker image in clean just adds more overhead to the
next cperf run, because the entire docker image has to be re-downloaded.
Docker pull will automatically download the difference between the old
image and the new image (less bandwidth) so there is no reason to remove
the image.

Change-Id: I79ba2af20e3bdacfedc28d468fb629e040d7098f
Signed-off-by: Tim Rozet <trozet@redhat.com>