Update patch set 3
Patch Set 3: Code-Review-2

Upgrading global-jjb and upgrading JJB should be two separate commits.
That way we can easily roll back if there is an issue with one or the other.

Given the majority of this patch is aimed at updating jjb, the global-jjb change should be pulled out.

Once we update to a version of jjb that fixes the 'block-level' we can switch the config back from 'blocking-level'.

Patch-set: 3
Label: Code-Review=-2