Call prepare-odl-tests playbook
Functest is running odl tests when the scenario contains odl. To run
those tests, functest requires a few variables which we can fetch
through the prepare-odl-tests.yml playbook.

Unfortunately:

- we cannot fetch the env variables from xci.env because the path to
PDF and IDF files is different. That's why we need to export them when
executing prepare-odl-tests.yml

- we cannot add the only task in prepare-odl-tests.yml to
prepare-tests.yml because they use a different inventory and we cannot
combine them because the odl inventory can only be executed from the
opnfv vm and prepare-tests.yml executs in the deployment host

Change-Id: I38f4399bc8c7bc5fbba0126a4b3e2d6970b4a32b
Signed-off-by: Manuel Buil <mbuil@suse.com>