Ensure jobs don't get queued on offline builders
A change to the NodeLabelParameter config in JJB caused any jobs using
the plugin to possibly have their jobs triggered on builders which
weren't online.

This in turn caused the queue to backup and newer jobs are blocked
waiting for older ones to complete, which never happens as they're never
queued on an online builder.

The 'all-nodes' parameter ensure jobs that are set to run concurrently
don't trigger across all available nodes at once, as this will include
offline nodes as well. As this is the default of the plugin, projects
which need it can enable it individually.

Change-Id: Ia690eef078209b6b056dad85613dda6868b18271
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>