Functest: Sets max concurrent jobs to 10
In Apex, some random jobs are queueing to intel-pod11 which does not
exist anymore. The problem is, when these jobs queue they are blocking
apex daily functest, as it looks like the job is only set to allow 1 job
globally at a time. This change bumps the max concurrent global jobs to
10.

Change-Id: I62d458a92abfa35bb0d1c459ccf4ffedfd51e66d
Signed-off-by: Tim Rozet <trozet@redhat.com>