Update patch set 1
Patch Set 1:

I should have been clear - as we do not have a formal process defined for CI/CD, I think it makes the most sense to continue with the regular release process.  Therefore this branch point is good and I would like to proceed with the creation of gambia and will let David know I'll be part of the final release.

Patch-set: 1