Apex: Adds ability to force snapshot promotion
We want the ability to be able to force a promotion and skip CSIT. This
patch introduces a variable we can use when manually triggering a
promotion job to force promotion. The use case here is for HA scenarios
we have no snapshots because CSIT always fails. We want to be able to
force promotions for this case to at least have some snapshot available.

Change-Id: I7e81b458f4169596929f032e9394695477fbb940
Signed-off-by: Tim Rozet <trozet@redhat.com>