Apex: Do not use apex branch in promote gs path
When we promote artifacts we do not want to include the apex branch (for
example "gambia") in the gs path URL for where snapshots are uploaded.
We may use a different apex branch to promote older OS versions of
snapshots, but we still want a unified url of just
apex/<os_version>/<topology> for our snapshot artifacts.

Change-Id: I7dfe40558ce05d04e1805884cd6a47c82220581e
Signed-off-by: Tim Rozet <trozet@redhat.com>