Write tempest_conf.yaml for all OPNFV installers
It follows "Move tempest logics in tempest_conf.yaml" [1] and then
allows passing tempest_full and tempest_scenario in OPNFV gates.

It set all params for all OPNFV installers according to their last
runs [2][3][4].

Live migration is out of that custom params as it's currently
considered as a key feature [5].

tempest_conf.yaml should be installer-centric and shouldn't take
scenarios into account.

It's unclear which installers support glance API v1 which is
deprecated. Any installer can ask for testing it if supported.
At least Apex doesn't support it [6].

Shelve is currenlty disabled because testing it could fail the
whole tempest run if unsupported. We are missing data about which
installer may support it.

[1] https://gerrit.opnfv.org/gerrit/#/c/64067
[2] Apex: https://artifacts.opnfv.org/logs/functest/lf-pod1/2018-10-25_02-17-25/tempest_full/tempest-report.html
[3] Compass: https://artifacts.opnfv.org/logs/functest/huawei-virtual9/2018-10-24_04-54-25/tempest_full/tempest-report.html
[4] Fuel: https://artifacts.opnfv.org/logs/functest/lf-pod2/gambia/2018-10-25_05-57-54/tempest_full/tempest-report.html
[5] https://artifacts.opnfv.org/logs/functest/ericsson-pod1/gambia/2018-10-27_06-10-07/tempest_full/tempest-report.html
[6] https://artifacts.opnfv.org/logs/functest/lf-pod1/gambia/2018-10-23_10-34-07/tempest_full/tempest-report.html

Change-Id: I17c8a2d738daaf1c2e97755c99f8886da8a531aa
Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>