Verify Functest on functest-virtual
It runs most of healthcheck and tempest-based test cases as a first
step. All benchmarking tools and all vnfs are not executed yet.

All active Functest versions are verified thanks to backward
compatibility ensured by Tempest and then by Functest.

The jobs are designed to be run in parallel (we could also run all
Functest versions in parallel instead of separated steps).

It will be switched to a verify gate in a second step.

Change-Id: I07cada848dd90f70da8741269692de8646813bdb
Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>