Verify Apex snapshots on functest-virtual
It runs most of tempest-based test cases as supported by the current
Apex snapshots.

All benchmarking tools and all vnfs and not executed (missing
services, lack of resources, etc.).

All active Functest versions are verified thanks to backward
compatibility ensured by Tempest and then by Functest.

The jobs are designed to be run in parallel (we could also run all
Functest versions in parallel instead of separated steps).

It will be switched to a verify gate in a second step.

Change-Id: I07cada848dd90f70da8741269692de8646813bdb
Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>