Apex: Fix OS_VERSION detection
When we run deployments for snapshot creation we typically pass in the
OS_VERSION var so we know what openstack version we are deploying with.
However when we run a verify/gate job to ensure that the snapshot
deploys work we do not have this variable, and need to detect it based
on the OPNFV branch, or if the branch is master, the deploy scenario
naming scheme.

Change-Id: If1445dc564d2af59b9740391de56b821b78de0b5
Signed-off-by: Tim Rozet <trozet@redhat.com>