Avoid possible deadblocks in Functest jjbs
From the time being, only 2 executors are allowed and then it may lead
to deadblocks if two multijobs are executed.

Normally we should only forbid gate and daily jobs at the time (tags are hardcoded).

It reverts to new patchset triggers as on core approval is buggy.

Daily jobs are now triggered once a day.

Change-Id: Iaa1ae73f8baeed037dd2f20f05ec3b15d1643b3b
Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>