Update patch set 1
Patch Set 1:

I'm testing a few possible workarounds for this, but so far I'm not happy with either of them.
The problem is indeed caused by iptables rules added by Docker for its networks that use the 'bridge' driver and are by default isolated from each other.
However, the iptables tables used by Docker changed a lot recently and since most of our CI hosts and possibly users are on Docker 17.05, we can't rely solely on DOCKER-USER table which was added in 17.06.
Also, injecting ALLOW rules in the FORWARD table is not clean and will lead to the same issue we had before with custom rules we kept on duplicating ...
I'll continue looking into this.

Patch-set: 1
Reviewer: Gerrit User 2997 <2997@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=0