[xci] Add installer type to scenario compatibility check
A check is made if a particular scenario is supported in a distro
and the scenario is skipped if not.

The check is extended to check for the installer type support as
well, for cases that a scenario is supported by more than one
installers each one with its own supported distros list.

Change-Id: Ic30ca7ba544a575ff01a75def8eed909dbcaa7b2
Signed-off-by: nikoskarandreas <nick@intracom-telecom.com>