Get XCI_FLAVOR as parameter from the commit msg
When a new patch is pushed on releng-xci repo and the impact
is not related to XCI_FLAVOR=mini, the environment preparation of CI
should be done, choosing as XCI_FLAVOR noha or ha.

This patch allows to get the XCI_FLAVOR parameter from commit message
and specifically from xci-flavor parameter.

Change-Id: Id5215e3fab1fc308435781204a6ed721a9c5a0c7
Signed-off-by: Panagiotis Karalis <pkaralis@intracom-telecom.com>