xci: Fix scenario determination logic
Function override_xci_flavor exits with 0 early, preventing
the rest of the script and especially override_scenario function
to run. This makes it impossible to verify patches that contains
installer-type or deploy-scenario in commit message since the
generated scenario.properties file contains only XCI_FLAVOR in
it, causing verify jobs to fail determining the installer type
and the scenario. See the example.

https://build.opnfv.org/ci/job/xci-verify-ubuntu-virtual-master/2937/console

This change removes exit 0 from override_xci_flavor and moves
the call to it to after override_scenario function.

JIRA: XCI-21

Change-Id: Ic7f943ba5e1562177b59a8215cd8a6f2a58db093
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>