Publish first CNTT jobs
It's based on Airship installer and the next intel-pod10 slave.

For the time, it selects the same Functest containers but they will be
updated according to the next CNTT doc updates and then the related
Functest changes (e.g. new functest-cntt-smoke container).

Both Airship and CNTT verification jobs will be triggered only
depending on the file modified.

The deployment calls may be updated in next changes (see
airship-deploy builder).

Change-Id: Ia330f3d61f87d143819693de09001f746e253e8d
Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>