Run smoke in parallel back
It's no simple to get the best threads (test cases and tempest) values
and then guarantee Functest vs:
  - test flapping
  - backend limitations

As opposed to the OpenStack gates, Jenkins runs all test cases in
parallel which indirectly increases the number of threads.

The SUT has been improved regarding the workers (divided by 2).
The default value was too high due to the number of cpus.

Change-Id: I09be861b9ec369f36447ac9e9c09d14f2b7da1a8
Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>