Update patch set 1
Patch Set 1:

I am happy to switch to whatever means of using Python3.

The idea of this verification job was not just to test if the build succeeds, but to render everything to one single file and upload it for review by a human. So the job used sed to overwrite the output format.

Patch-set: 1