Get rid of global jjb in Anuket
They are useless here, never updated and failed
a lot because of package management issues in it.

It now defines all views in the classical way.

It also removes all rtd jobs as they are useless.
Our docs are built on git events via RTD too.
(and it avoids the hardcoded token which should be
a security convern).

If needed, cirv and laas python verification will
be updated in a second patchset.

Change-Id: I5aee179d804a5547a464501a03d7afa395420552
Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>